refactor: 로또 생성 전략 패키지 변경
- 이전: utils
- 이후: model/generator
