+++ src/main/kotlin/model/generator/LotteriesGenerationStrategy.kt
@@ -1,4 +1,4 @@
-package utils
+package model.generator
 
 import model.Lottery
