+++ src/test/kotlin/model/generator/ExplicitLotteriesGenerationStrategy.kt
@@ -1,4 +1,4 @@
-package utils
+package model.generator
 
 import model.Lottery
